+++ backend/src/histories/histories.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { HistoriesService } from 'src/service/histories.service';
-import { historiesProviders } from '../providers/histories.providers'; // Import the 'historiesProviders' variable from the correct file
+import { HistoriesService } from 'src/histories/histories.service';
+import { historiesProviders } from './histories.providers'; // Import the 'historiesProviders' variable from the correct file
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { VHistories } from 'src/entities/VHistories';
-import { HistoriesController } from 'src/controller/histories.controller';
+import { HistoriesController } from 'src/histories/histories.controller';
 
 @Module({
   imports: [TypeOrmModule.forFeature([VHistories])],
